% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateComponents.R
\name{CreateComponents}
\alias{CreateComponents}
\title{CreateComponents}
\usage{
CreateComponents(
  fixedEffects,
  likeLihoodIntercept = TRUE,
  speciesIntercept = FALSE,
  speciesInterceptModel = list(),
  speciesSpatial = TRUE,
  speciesSpatialModel = list(),
  biasSpatial = FALSE,
  customComponents,
  fixedPriors = c(prior.mean = 0, prior.prec = 1)
)
}
\arguments{
\item{fixedEffects}{A vector of the names of the fixed effects included in the model.}

\item{speciesIntercept}{Create an intercept for the species. Defaults to \code{FALSE}.}

\item{speciesInterceptModel}{The model for the species intercept. Defaults to xxx.}

\item{speciesSpatial}{Create a pecies spatial effect included in the model. Defaults to \code{TRUE}.}

\item{speciesSpatialModel}{The model for the species spatial. Defaults to xxx.}

\item{biasSpatial}{Include a bias spatial effect in the model (logical).}

\item{customComponents}{Custom components to include in the model.}

\item{fixedPriors}{Priors for the fixed (environmental and observation effects). Must be a named vector containing: \code{prior.mean} and \code{prior.prec}.}

\item{likelihoodIntercept}{Create an intercept for each likelihood. Defaults to \code{TRUE}.}
}
\description{
Create components required by \pkg{inlabru}.
}
