% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateIPS.R
\name{CreateIPS}
\alias{CreateIPS}
\title{CreateIPS}
\usage{
CreateIPS(graph, ...)
}
\arguments{
\item{graph}{A graph from the \pkg{MetricGraph} package.}

\item{...}{Additional arguments to be used by \code{.$build_mesh} (if a mesh has not been provided).}
}
\value{
A \pkg{sf} \textit{POINTS} object containing the IPS with the needed information.
}
\description{
Create integration points and add them to the model.
}
