% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractCovariates.R
\name{ExtractCovariates}
\alias{ExtractCovariates}
\title{extractCovariates}
\usage{
ExtractCovariates(data, covariates, used = NULL)
}
\arguments{
\item{data}{A \code{metric_graph_data} object.}

\item{covariates}{A sf object of covariate data.}

\item{used}{Which spatial covariates to include in the model.}
}
\value{
A data.frame object containing the covariate information.
}
\description{
Assign covariate information to the observation data and integration points
}
