% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLikelihoods.R
\name{MakeLikelihoods}
\alias{MakeLikelihoods}
\title{MakeLikelihoods}
\usage{
MakeLikelihoods(
  data,
  components,
  effectsList = list(PO = NULL, PA = NULL, Counts = NULL),
  formulaList = list(PO = y ~ ., PA = y ~ ., Counts = y ~ .),
  likelihoodFamily = list(PO = "poisson", PA = "binomial", Counts = "poisson"),
  predictionFormula = NULL
)
}
\arguments{
\item{data}{A \code{metric_graph_data} object.}

\item{components}{Components to include from the model. Must be the output of \code{\link{CreateComponents}}.}

\item{effectsList}{A named list containing the vectors of named effects for each likelihood. If \code{NULL} for a likelihood, will use all effects in components.}

\item{formulaList}{A named list containing the formulas for each likelihood.}

\item{likelihoodFamily}{A named list of the statistical families for each likelihood.}

\item{predictionFormula}{A named list of formulas for predicting on. Defaults to \code{NULL} which does not create any prediction formulas. The response variable for the formula needs to be y}
}
\value{
A \link[inlabru]{bru_obs_list} object containing the likelihoods to be used by \link[inlabru]{bru}.
}
\description{
Create a bru_obs_list object of likelihoods for each dataset.
}
